Add set_device_map to TensorPipeOptions to support GPU args
This commit enables sending non-CPU tensors through RPC using
TensorPipe backend. Users can configure device mappings by calling
`set_device_map` on `TensorPipeRpcBackendOptions`. Internally,
the `init_rpc` API verifies the correctness of device mappings. It
will shutdown RPC if the check failed, or proceed and pass global
mappings to TensorPipeAgent if the check was successful. For serde,
we added a device indices field to TensorPipe read and write buffers,
which should either be empty (all tensors must be on CPU) or match
the tensors in order and number in the RPC message. This commit
does not yet implement zero-copy, the tensor is always moved to CPU
on the sender and then moved to the specified device on the receiver.

ghstack-source-id: 64afd23479315084ce2cc95624af901dc9d4ba8f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/42637